+++ src/gen/genModule.test.ts
@@ -1,34 +1,34 @@
 import {suite} from 'uvu';
-import * as t from 'uvu/assert';
+import * as assert from 'uvu/assert';
 import {join} from 'path';
 
 import {validateGenModule, findGenModules} from './genModule.js';
 import {paths} from '../paths.js';
 import {fs} from '../fs/node.js';
 
-/* testValidateGenModule */
-const testValidateGenModule = suite('validateGenModule');
+/* test__validateGenModule */
+const test__validateGenModule = suite('validateGenModule');
 
-testValidateGenModule('minimal interface', () => {
-	t.ok(validateGenModule({gen: () => {}}));
+test__validateGenModule('minimal interface', () => {
+	assert.ok(validateGenModule({gen: () => {}}));
 });
 
-testValidateGenModule('invalid module', () => {
-	t.not.ok(validateGenModule({gen: {}}));
-	t.not.ok(validateGenModule({task: {run: {}}}));
+test__validateGenModule('invalid module', () => {
+	assert.not.ok(validateGenModule({gen: {}}));
+	assert.not.ok(validateGenModule({task: {run: {}}}));
 });
 
-testValidateGenModule.run();
-/* /testValidateGenModule */
+test__validateGenModule.run();
+/* test__validateGenModule */
 
-/* testFindGenModules */
-const testFindGenModules = suite('findGenModules');
+/* test__findGenModules */
+const test__findGenModules = suite('findGenModules');
 
-testFindGenModules('finds gen modules in a directory', async () => {
+test__findGenModules('finds gen modules in a directory', async () => {
 	const findGenModulesResult = await findGenModules(fs, [join(paths.source, 'docs/')]);
-	t.ok(findGenModulesResult.ok);
-	t.ok(findGenModulesResult.sourceIdPathDataByInputPath.size);
+	assert.ok(findGenModulesResult.ok);
+	assert.ok(findGenModulesResult.sourceIdPathDataByInputPath.size);
 });
 
-testFindGenModules.run();
+test__findGenModules.run();
 /* /findGenModulesResult */