+++ src/utils/gitignore.test.ts
@@ -1,37 +1,37 @@
 import {resolve} from 'path';
 import {suite} from 'uvu';
-import * as t from 'uvu/assert';
+import * as assert from 'uvu/assert';
 
 import {loadGitignoreFilter} from './gitignore.js';
 
-/* testLoadGitignoreFilter */
-const testLoadGitignoreFilter = suite('loadGitignoreFilter');
+/* test__loadGitignoreFilter */
+const test__loadGitignoreFilter = suite('loadGitignoreFilter');
 
-testLoadGitignoreFilter('basic behavior', () => {
+test__loadGitignoreFilter('basic behavior', () => {
 	const filter = loadGitignoreFilter();
-	t.ok(filter(resolve('dist')));
-	t.ok(!filter(resolve('a/dist')));
-	t.ok(filter(resolve('node_modules')));
-	t.ok(filter(resolve('a/node_modules')));
-	t.ok(filter(resolve('node_modules/a')));
-	t.ok(filter(resolve('a/node_modules/b')));
-	t.ok(!filter(resolve('node_module')));
-	t.ok(!filter(resolve('a/node_module')));
-	t.ok(!filter(resolve('node_module/a')));
-	t.ok(!filter(resolve('a/node_module/b')));
+	assert.ok(filter(resolve('dist')));
+	assert.ok(!filter(resolve('a/dist')));
+	assert.ok(filter(resolve('node_modules')));
+	assert.ok(filter(resolve('a/node_modules')));
+	assert.ok(filter(resolve('node_modules/a')));
+	assert.ok(filter(resolve('a/node_modules/b')));
+	assert.ok(!filter(resolve('node_module')));
+	assert.ok(!filter(resolve('a/node_module')));
+	assert.ok(!filter(resolve('node_module/a')));
+	assert.ok(!filter(resolve('a/node_module/b')));
 });
 
-testLoadGitignoreFilter('caching and forceRefresh', () => {
+test__loadGitignoreFilter('caching and forceRefresh', () => {
 	const filter1 = loadGitignoreFilter();
 	const filter2 = loadGitignoreFilter();
-	t.is(filter1, filter2);
+	assert.is(filter1, filter2);
 	const filter3 = loadGitignoreFilter(true);
-	t.is.not(filter1, filter3);
+	assert.is.not(filter1, filter3);
 	const filter4 = loadGitignoreFilter(false);
-	t.is(filter3, filter4);
+	assert.is(filter3, filter4);
 	const filter5 = loadGitignoreFilter(true);
-	t.is.not(filter4, filter5);
+	assert.is.not(filter4, filter5);
 });
 
-testLoadGitignoreFilter.run();
-/* /testLoadGitignoreFilter */
+test__loadGitignoreFilter.run();
+/* test__loadGitignoreFilter */