+++ svelte.config.js
@@ -1,9 +1,28 @@
 import {typescript} from 'svelte-preprocess-esbuild';
+import staticAdapter from '@sveltejs/adapter-static';
 
 /** @type {import('@sveltejs/kit').Config} */
 export default {
 	preprocess: typescript(),
 	compilerOptions: {
 		immutable: true,
+	},
+	kit: {
+		adapter: staticAdapter(),
+		target: '#root',
+		files: {assets: 'src/static'},
+		vite: {
+			server: {
+				proxy: {
+					'/api': 'http://localhost:8999',
+				},
+			},
+			ssr: {
+				noExternal: ['@feltcoop/felt'],
+			},
+			optimizeDeps: {
+				exclude: ['@feltcoop/felt'],
+			},
+		},
 	},
 };